+++ package.json
@@ -7,6 +7,7 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "immer": "^10.0.2",
     "json-server": "^0.17.3",
     "nanoid": "^4.0.2",
     "react": "^18.2.0",
